Changes to tour and set up
I am still in the process of working on this, and am getting an error
about the scavengerHuntIsSetUp variable being nil, but I don’t know
exactly where to change it. I don’t know if this a small thing Jared
might have changed when working on the data handling or something I
changed today. Thoughts?